refactor: 스터디 장의 id의 타입을 Long으로 변경

+++ android/app/src/main/java/com/created/team201/data/remote/response/StudyDetailResponseDto.kt
@@ -16,7 +16,7 @@ data class StudyDetailResponseDto(
     @SerialName("numberOfMaximumMembers")
     val numberOfMaximumMembers: Int,
     @SerialName("studyMasterId")
-    val studyMasterId: Int,
+    val studyMasterId: Long,
     @SerialName("role")
     val role: Int,
     @SerialName("startAt")
